Fix SNR-datagen
- Save as float16 instead of float64... storage savings!!!
- Save object masks to disk

+++ classes/slice_class.py
@@ -1,3 +1,6 @@
+import numpy as np
+
+
 class SliceObj:
     def __init__(self, arr):
-        self.data = arr
+        self.data = arr.astype(np.float16)
